baikal-t/support-requests#13
Micropython was present in the old sumo build, so we include it for
honister too. Use rather minimal micropython setup, withtout
FROZEN_MANIFEST and MPYCROSS functionality.  This is similar to
micropython-minimal build target. If MPYCROSS is required, figure out
how to build it natively, without the CROSS_PREFIX stuff in
EXTRA_OEMAKE. Also, use semicolon instead of underscore (new syntax).